Extract atomic file write utility to shared fileutils package
The atomicWriteFile function was previously defined in pkg/workloads/statuses/file_status.go
and used only for status file writes. This refactoring extracts it to a new pkg/fileutils package
as AtomicWriteFile, making it reusable across the codebase.

Changes:
- Create new pkg/fileutils package with AtomicWriteFile function
- Move atomic write tests to pkg/fileutils/atomic_test.go
- Update pkg/client/config_editor.go to use fileutils.AtomicWriteFile for JSON/YAML config writes
- Update pkg/workloads/statuses/file_status.go to use fileutils.AtomicWriteFile for status writes
- Remove duplicate atomicWriteFile implementation and tests from file_status.go

This improves code reusability and ensures consistent atomic write behavior across all file operations
(config files, status files) preventing partial writes during concurrent access or system failures.
