+++ pkg/workloads/statuses/file_status_test.go
@@ -2014,93 +2014,6 @@ func TestFileStatusManager_IsRemoteWorkload_EdgeCases(t *testing.T) {
 	}
 }
 
-func TestAtomicWriteFile(t *testing.T) {
-	t.Parallel()
-
-	tempDir := t.TempDir()
-
-	tests := []struct {
-		name        string
-		data        []byte
-		perm        os.FileMode
-		expectError bool
-	}{
-		{
-			name:        "successful write",
-			data:        []byte(`{"test": "data"}`),
-			perm:        0o600,
-			expectError: false,
-		},
-		{
-			name:        "empty data",
-			data:        []byte{},
-			perm:        0o600,
-			expectError: false,
-		},
-		{
-			name:        "large data",
-			data:        []byte(strings.Repeat("x", 10000)),
-			perm:        0o644,
-			expectError: false,
-		},
-	}
-
-	for _, tt := range tests {
-		tt := tt
-		t.Run(tt.name, func(t *testing.T) {
-			t.Parallel()
-			// Use different file for each test to avoid conflicts
-			testPath := filepath.Join(tempDir, tt.name+".json")
-
-			err := atomicWriteFile(testPath, tt.data, tt.perm)
-
-			if tt.expectError {
-				assert.Error(t, err)
-			} else {
-				assert.NoError(t, err)
-
-				// Verify file exists and has correct content
-				content, readErr := os.ReadFile(testPath)
-				require.NoError(t, readErr)
-				assert.Equal(t, tt.data, content)
-
-				// Verify permissions
-				info, statErr := os.Stat(testPath)
-				require.NoError(t, statErr)
-				assert.Equal(t, tt.perm, info.Mode().Perm())
-			}
-		})
-	}
-}
-
-func TestAtomicWriteFile_Overwrite(t *testing.T) {
-	t.Parallel()
-
-	tempDir := t.TempDir()
-	targetPath := filepath.Join(tempDir, "test.json")
-
-	// Write initial data
-	initialData := []byte(`{"initial": "data with more content to ensure truncation"}`)
-	err := atomicWriteFile(targetPath, initialData, 0o600)
-	require.NoError(t, err)
-
-	// Verify initial write
-	content, err := os.ReadFile(targetPath)
-	require.NoError(t, err)
-	assert.Equal(t, initialData, content)
-
-	// Overwrite with smaller data
-	newData := []byte(`{"new": "data"}`)
-	err = atomicWriteFile(targetPath, newData, 0o600)
-	require.NoError(t, err)
-
-	// Verify overwrite - should be only the new data, not appended
-	content, err = os.ReadFile(targetPath)
-	require.NoError(t, err)
-	assert.Equal(t, newData, content)
-	assert.Len(t, content, len(newData), "file should be truncated to new data size")
-}
-
 func TestJSONRecovery_ExtraClosingBrace(t *testing.T) {
 	t.Parallel()
